% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{align_images}
\alias{align_images}
\title{Align two images}
\usage{
align_images(img1, img2)
}
\arguments{
\item{img1}{image 1}

\item{img2}{image 2}
}
\value{
an image list with image 1 and image 2 on the same coordinate system
}
\description{
Align two images
}
