% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{apply_homograph}
\alias{apply_homograph}
\title{Apply homography to points in p}
\usage{
apply_homograph(h, p)
}
\arguments{
\item{h}{homography}

\item{p}{points}
}
\description{
Source: https://www.kaggle.com/vicensgaitan/image-registration-the-r-way/notebook
}
