% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{contour_ellipse_fit}
\alias{contour_ellipse_fit}
\title{Fit an ellipse to the boundary contour points}
\usage{
contour_ellipse_fit(contour_points, chull = F)
}
\arguments{
\item{contour_points}{a cimg or data frame of contour points.
If a cimg, must be a pixset (0s and 1s). If a data frame, must have
columns x and y}

\item{chull}{use the convex hull of the points?}
}
\value{
a data frame of ellipse parameters: X, Y for the center, A, B for the
          major and minor axes, and angle for the rotation angle.
}
\description{
Fit an ellipse to the boundary contour points
}
