% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{descriptor_orientation}
\alias{descriptor_orientation}
\title{Calculate orientation of keypoint descriptors}
\usage{
descriptor_orientation(im, stencil_ext = NULL, stencil = NULL, theta,
  v)
}
\arguments{
\item{im}{image}

\item{stencil_ext}{stencil used to get a patch around the keypoint}

\item{stencil}{stencil used to calculate the descriptor}

\item{theta}{angle to rotate the image}

\item{v}{numeric vector of length 2 giving x and y coordinates of the keypoint}
}
\description{
Source: https://www.kaggle.com/vicensgaitan/image-registration-the-r-way/notebook
}
