% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_check}
\alias{ellipse_check}
\title{Assertions to check that an ellipse object has the correct form}
\usage{
ellipse_check(ellipse)
}
\arguments{
\item{ellipse}{object from contour_ellipse_fit, or a data frame with columns
CenterX, CenterY, AxisA, AxisB, and Angle}
}
\value{
TRUE if all are met, error messages otherwise
}
\description{
Assertions to check that an ellipse object has the correct form
}
