% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_points}
\alias{ellipse_points}
\title{Return the boundary points of an ellipse}
\usage{
ellipse_points(ellipse, n = 300, plot_lines = T, ...)
}
\arguments{
\item{ellipse}{output from contour_ellipse_fit, or a data frame with the
following columns: CenterX, CenterY, AxisA, AxisB, Angle}

\item{n}{number of points}

\item{plot_lines}{return call to lines?}

\item{...}{additional parameters for lines call}
}
\value{
two-column data frame of x and y points
}
\description{
Return the boundary points of an ellipse
}
