% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{est_homograph}
\alias{est_homograph}
\title{Estimate homography from points in P to points in p}
\usage{
est_homograph(P, p)
}
\arguments{
\item{P}{set of points}

\item{p}{set of points in different space}
}
\description{
Source: https://www.kaggle.com/vicensgaitan/image-registration-the-r-way/notebook
}
