% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{halfimg_split}
\alias{halfimg_split}
\title{Split an image in half along an axis, mirror the 2nd half of the image, and return the image list}
\usage{
halfimg_split(img, axis = "y")
}
\arguments{
\item{img}{a cimg}

\item{axis}{"x" or "y"}
}
\value{
an imlist
}
\description{
Split an image in half along an axis, mirror the 2nd half of the image, and return the image list
}
