% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{halfimg_unsplit}
\alias{halfimg_unsplit}
\title{Paste a split image back together, first mirroring the 2nd entry in the imlist}
\usage{
halfimg_unsplit(imlist, axis = "y")
}
\arguments{
\item{imlist}{an imlist}

\item{axis}{"x" or "y"}
}
\value{
an imlist
}
\description{
Paste a split image back together, first mirroring the 2nd entry in the imlist
}
