% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{harris_corners}
\alias{harris_corners}
\title{Harris Keypoint detector}
\usage{
harris_corners(im, sigma = 2, eps = 1e-10)
}
\arguments{
\item{im}{image}

\item{sigma}{sigma for gaussian blur function}

\item{eps}{threshold (defaults to 1e-10)}
}
\value{
corner function value
}
\description{
The value of the Harris transformation is high only in well defined corners
on the image.
}
\details{
Source: https://www.kaggle.com/vicensgaitan/image-registration-the-r-way/notebook
}
