% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{harris_keypoints}
\alias{harris_keypoints}
\title{Detect keypoints using Harris corner detector, and find centers of associated regions}
\usage{
harris_keypoints(im, thr = "99\%", sigma = 3, bord = 10)
}
\arguments{
\item{im}{image}

\item{thr}{threshold, as passed to imager::threshold}

\item{sigma}{blur used for corner detection}

\item{bord}{border region}
}
\value{
a list, with an image labeled by region and a data frame of the region's centers.
}
\description{
Runs the harris_corners function, then selects as keypoints the center of
connected regions with Harris over a certain threshod.
}
