% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{hog}
\alias{hog}
\title{Histogram of oriented gradients}
\usage{
hog(im)
}
\arguments{
\item{im}{image}
}
\value{
a list of one or two critical orientations
}
\description{
Histogram of oriented gradients
}
