% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShoeAlignR-package.R
\docType{package}
\name{ImageAlignR-package}
\alias{ImageAlignR}
\alias{ImageAlignR-package}
\title{ImageAlignR: ImageAlignR}
\description{
RANSAC, ellipse-based alignment, and more.
}
\author{
\strong{Maintainer}: Susan Vanderplas \email{srvander@iastate.edu}

}
\keyword{internal}
