% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{images_resize}
\alias{images_resize}
\title{Function to make two images the same size by padding one image}
\usage{
images_resize(img1, img2)
}
\arguments{
\item{img1}{image 1}

\item{img2}{image 2}
}
\description{
Function to make two images the same size by padding one image
}
