% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.R
\name{img_check}
\alias{img_check}
\title{Check image to ensure it has the right type}
\usage{
img_check(img, keep_alpha = F, keep_color = F)
}
\arguments{
\item{img}{an object to check to see if it is a cimg with the right channels}

\item{keep_alpha}{should the image keep an alpha channel?}

\item{keep_color}{should the image be in color?}
}
\value{
a cimg of the correct specifications (if possible), otherwise, will
          error out
}
\description{
Check image to ensure it has the right type
}
