% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{knn_points}
\alias{knn_points}
\title{Get KNN for feature point matches}
\usage{
knn_points(
  kpf_a,
  kpf_b,
  centers_a,
  centers_b,
  n = 2,
  ratio = 0.8,
  show_plot = F
)
}
\arguments{
\item{kpf_a}{matrix of features for image a}

\item{kpf_b}{matrix of features for image b}

\item{centers_a}{matrix of points for image a}

\item{centers_b}{matrix of points for image b}

\item{n}{number of neighbors to keep}

\item{ratio}{ratio of first neighbor to second neighbor distance. Values smaller than this will be kept}

\item{show_plot}{show a plot of matches kept?}
}
\description{
Source: https://www.kaggle.com/vicensgaitan/image-registration-the-r-way/notebook
}
