% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{map_affine_gen}
\alias{map_affine_gen}
\title{Function to generate an affine transform}
\usage{
map_affine_gen(homography)
}
\arguments{
\item{homography}{homograph matrix from ransac (unsolved)}
}
\description{
Returns a function taking arguments x and y, for use in imager::imwarp
}
