% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_halfimg}
\alias{map_halfimg}
\title{Split an image in half along one or more axes, apply a function, and paste it back together.}
\usage{
map_halfimg(img, fun = function(x) x, axes = "xy")
}
\arguments{
\item{img}{a cimg object}

\item{fun}{a function that operates on a cimg}

\item{axes}{"x", "y", or "xy"}
}
\value{
an imlist
}
\description{
Split an image in half along one or more axes, apply a function, and paste it back together.
}
