% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{oriented_gradients}
\alias{oriented_gradients}
\title{Find most common gradients}
\usage{
oriented_gradients(im, sigma = 0, show_plot = T)
}
\arguments{
\item{im}{image}

\item{sigma}{blur to use for hog image}

\item{show_plot}{show histogram of gradients?}
}
\value{
a list of one or two critical orientations
}
\description{
Source: https://www.kaggle.com/vicensgaitan/image-registration-the-r-way/notebook
}
