% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.R
\name{outer_contour}
\alias{outer_contour}
\title{Get outer contour of the object}
\usage{
outer_contour(img, thr = mean(img), as_cimg = TRUE)
}
\arguments{
\item{img}{a cimg object}

\item{thr}{threshold to use for the image}

\item{as_cimg}{return the points as a cimg object?}
}
\value{
either a cimg of points or a data frame of points indicating the boundary
}
\description{
This function scans each row and column of the image and finds the minimum
and maximum coordinates which are less than thr (by default, the mean value
of the image).
}
