% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantize_colors}
\alias{quantize_colors}
\title{Limit the number of colors in the image}
\usage{
quantize_colors(img, n = 16)
}
\arguments{
\item{img}{cimage}

\item{n}{number of colors in returned image}
}
\value{
a cimage with only n colors
}
\description{
From https://lumiamitie.github.io/r/imager-color-quantization/
}
