% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{ransac}
\alias{ransac}
\title{Robust homography estimation from p1 to p2.}
\usage{
ransac(p1, p2, thresh = 100, N = 1000)
}
\arguments{
\item{p1}{point set}

\item{p2}{point set}

\item{thresh}{threshold value}

\item{N}{samples to use}
}
\description{
Source: https://www.kaggle.com/vicensgaitan/image-registration-the-r-way/notebook
}
\details{
Returns h and the list of inliers
}
