% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harris.R
\name{region_centers}
\alias{region_centers}
\title{Get centers of a set of labeled regions within an image}
\usage{
region_centers(im, bord = 10)
}
\arguments{
\item{bord}{border size}
}
\value{
a data frame of coordinates
}
\description{
Get centers of a set of labeled regions within an image
}
