% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShoeAlignR-package.R
\docType{package}
\name{ShoeAlignR-package}
\alias{ShoeAlignR}
\alias{ShoeAlignR-package}
\title{ShoeAlignR: Aligning images (shoeprints in particular)}
\description{
RANSAC, ellipse-based alignment, and more.
}
\author{
\strong{Maintainer}: Susan Vanderplas \email{srvander@iastate.edu}

}
\keyword{internal}
