% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.R
\name{thin_contour}
\alias{thin_contour}
\title{Thin outer contour by removing points overlappling radially from the centroid}
\usage{
thin_contour(contour, img = NULL, centroid = NULL, n_angles = 1800,
  as_cimg = TRUE)
}
\arguments{
\item{contour}{results from outer_contour}

\item{img}{thresholded image (used to calculate the centroid)}

\item{centroid}{centroid of the image}

\item{n_angles}{number of unique angles to use}

\item{as_cimg}{return the points as a cimg object?}
}
\value{
either a cimg of points or a data frame of points indicating the 
  boundary
}
\description{
Thin outer contour by removing points overlappling radially from the centroid
}
