% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_and_refit.R
\name{img_rotate_refit}
\alias{img_rotate_refit}
\title{Rotate the image and re-fit the bounding ellipse}
\usage{
img_rotate_refit(img, ellipse, ...)
}
\arguments{
\item{img}{a cimg of the object to be rotated}

\item{ellipse}{the original (potentially tilted) ellipse}

\item{...}{additional parameters to pass to imrotate}
}
\description{
This should ensure that the major axis of the ellipse is vertical.
}
